requirements.txt
streamlit
torch
opencv-python-headless
Pillow
